requirements.txt
Flask
gunicorn
keras
numpy
Pillow
tensorflow